pyproject.toml
[project]
name = "dq-swirl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Jae Ro", email = "jaero.gatech@gmail.com" }
]

requires-python = ">=3.14"
dependencies = [
    "boto3>=1.42.37",
    "httpx>=0.28.1,<1.0.01",
    "pydantic>=2.12.5,<3.0.0",
    "python-dotenv>=1.2.1,<2.0.0",
    "redis>=7.1.0,<8.0.0",
    "saq>=0.26.1,<1.0.0",
    "virt-s3>=0.1.5",
]

[dependency-groups]
dev = [
    "ruff",
    "aiohttp-basicauth"
]
test = [
    "pytest",
    "pytest-asyncio",
]

[project.optional-dependencies]
app = [
    "quart>=0.20.0,<1.0.0",
    "quart-cors>=0.8.0,<1.0.0",
    "hypercorn>=0.18.0,<1.0.0"
]

worker = [
    "langgraph>=1.0.7,<2.0.0",
    "litellm>=1.81.4,<2.0.0",
]

[project.scripts]
swirl = "dq_swirl:hello"

[tool.uv]
default-groups = []

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.ruff]
exclude = [
    ".git",
    ".eggs",
    ".ipynb-checkpoints", 
    ".pytype", 
    ".pyenv",
    ".venv", 
    ".vscode", 
    "__pypackages__",
    "node_modules",
    "build", 
    "dist",
    "site-packages",
    ".ipynb",
    ".ruff_cache"
]

line-length = 88
indent-width = 4
target-version = "py314"

[tool.ruff.lint]
ignore = ["F401"]
fixable = ["ALL"]

# allow for unused variables when prefixed with underscore
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"


[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
